remove react query useLocalStorage
